botlog/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
defusedxml==0.7.1
diff-match-patch==20241021
distro==1.9.0
Django==5.1.3
django-admin-extra-urls==4.1.1
django-adminplus==0.6
django-environ==0.11.2
django-extensions==3.2.3
django-import-export==4.3.3
djangorestframework==3.15.2
et_xmlfile==2.0.0
h11==0.14.0
hiredis==3.1.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.7.1
msgpack-python==0.5.6
numpy==2.1.3
odfpy==1.4.1
openai==1.55.1
openpyxl==3.1.5
pandas==2.2.3
pydantic==2.10.1
pydantic_core==2.27.1
pymemcache==4.0.0
pyTelegramBotAPI==4.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jsend-response==0.0.3
python-telegram-bot==21.7
pytz==2024.2
PyYAML==6.0.2
redis==2.10.6
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.2
starlette==0.41.3
tablib==3.7.0
tabulate==0.9.0
telegraph==2.2.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
xlrd==2.0.1
xlwt==1.3.0
